add tests to parser; rename parser files
